Nå virker hibernate med Spring data jpa.

+++ eDrops/src/main/kotlin/com/sigmondsmart/edrops/domain/BlogEntry.kt
@@ -18,6 +18,6 @@ class BlogEntry(
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Int?=null
+    val id: Long?=null
 
 )
